feat(account): enhance event handling and validation for deposit and transfer workflows
- Updated event subscriptions in app.go to include logger and unit of work dependencies for deposit and transfer operations.
- Refactored event-driven handlers to improve clarity and maintainability, ensuring proper logging and error handling.
- Introduced new event types for money conversion and payment initiation, enhancing the event-driven architecture.
- Added detailed documentation for the event-driven deposit flow, clarifying the sequence of events and interactions.
- Removed obsolete code and tests related to previous event handling, streamlining the codebase.

+++ pkg/domain/account/events/transfer.go
@@ -11,7 +11,7 @@ type TransferRequestedEvent struct {
 	DestAccountID   uuid.UUID
 	SenderUserID    uuid.UUID
 	ReceiverUserID  uuid.UUID
-	Amount          int64
+	Amount          float64 // main unit (e.g., dollars)
 	Currency        string
 	Source          string // MoneySource as string
 	Timestamp       int64
